+++ sdk/src/main/java/ly/count/android/sdk/HealthTracker.java
@@ -1,0 +1,13 @@
+package ly.count.android.sdk;
+
+interface HealthTracker {
+    void logWarning();
+
+    void logError();
+
+    void logFailedNetworkRequest(int statusCode, String errorResponse);
+
+    void clearAndSave();
+
+    void saveState();
+}