+++ sdk/src/main/java/ly/count/android/sdk/RequestQueueProvider.java
@@ -67,4 +67,6 @@ interface RequestQueueProvider {
     String prepareFeedbackListRequest();
 
     String prepareServerConfigRequest();
+
+    String prepareHealthCheckRequest(String preparedMetrics);
 }